Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 
-float addition(int first, int second)
+int addition(int first, int second)
 {
     int counter= first + second;
     return counter;
 }
 
-float subtraction(int first, int second)
+int subtraction(int first, int second)
 {
     int counter= first - second;
     return counter;
 }
 
-float multiplication(int first, int second)
+int multiplication(int first, int second)
 {
     int counter= first * second;
     return counter;
 }
 
-float division(int first, int second)
+int division(int first, int second)
 {   if (second== 0)
     {
         return 0;
@@ -32,23 +32,23 @@ int main(void)
     int first, second;
     char special_character;
     printf("");
-    scanf("%f %f %c", &first, &second, &special_character);
+    scanf("%i %i %c", &first, &second, &special_character);
 
     if (special_character== '+')
     {
-        printf("%f", addition(first, second));
+        printf("%i", addition(first, second));
     }
     else if (special_character== '-')
     {
-        printf("%f", subtraction(first, second));
+        printf("%i", subtraction(first, second));
     }
     else if (special_character== '*')
     {
-        printf("%f", multiplication(first, second));
+        printf("%i", multiplication(first, second));
     }
     else
     {
-        printf("%f", division(first, second));
+        printf("%i", division(first, second));
     }
 
     return 0;
